Rename PanelLinksEditor to DrilldownLinksEditor and mote it to grafana/ui
